move status to a hook
