+++ src/hooks/useRumbleStatus/index.ts
@@ -1,0 +1,1 @@
+export { default as useRumbleStatus } from './useRumbleStatus';
