+++ src/hooks/useRumbleStatus/useRumbleStatus.ts
@@ -1,0 +1,21 @@
+import { useMemo } from 'react';
+import { Rumbles } from '../../api';
+
+const useRumbleStatus = (rumble: Rumbles.RumblePhases): string[] => {
+  const status = useMemo(() => {
+    switch (rumble) {
+      case 'ACTIVE':
+      case 'FEEDBACK':
+        return 'Active';
+      // TODO what should complete say
+      case 'COMPLETE':
+        return 'Closed';
+      case 'INACTIVE':
+      default:
+        return 'Scheduled';
+    }
+  }, [rumble]);
+  return [status];
+};
+
+export default useRumbleStatus;
